requirements.txt
boto3==1.26.60
pyjwt==2.6.0
